FIX: Color to rgb_matrix::Color

+++ include/modules/game-of-life/game_of_life_configuration.hpp
@@ -1,20 +1,18 @@
 #pragma once
 
-#include "modules/module_configuration.hpp"
 #include "common/models/color.hpp"
+#include "modules/module_configuration.hpp"
 
-namespace GameOfLife
-{
-  class Configuration : public ModuleConfiguration
-  {
-  public:
-    Configuration();
+namespace GameOfLife {
+class Configuration : public ModuleConfiguration {
+public:
+  Configuration();
 
-    virtual char *getHelp() const override;
-    virtual void parseArguments(char **argv, int argc) override;
+  virtual char *getHelp() const override;
+  virtual void parseArguments(char **argv, int argc) override;
 
-    int duration;
-    Color color;
-    bool generateColor;
-  };
-}
+  float duration;
+  Color color;
+  bool generateColor;
+};
+} // namespace GameOfLife
